rename hat file
