--- nemo/collections/asr/modules/hat.py
@@ -1,531 +1,0 @@
-# Copyright (c) 2022, NVIDIA CORPORATION.  All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# Copyright 2017 Johns Hopkins University (Shinji Watanabe)
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from typing import Any, Dict, List, Optional, Tuple, Union
-
-import torch
-from omegaconf import DictConfig, OmegaConf
-
-from nemo.collections.asr.modules import rnnt, rnnt_abstract
-from nemo.collections.asr.parts.submodules import stateless_net
-from nemo.collections.asr.parts.utils import adapter_utils, rnnt_utils
-from nemo.collections.common.parts import rnn
-from nemo.core.classes import adapter_mixins, typecheck
-from nemo.core.classes.exportable import Exportable
-from nemo.core.classes.mixins import AdapterModuleMixin
-from nemo.core.neural_types import (
-    AcousticEncodedRepresentation,
-    ElementType,
-    EmbeddedTextType,
-    LabelsType,
-    LengthsType,
-    LogprobsType,
-    LossType,
-    NeuralType,
-    SpectrogramType,
-)
-from nemo.utils import logging
-
-
-class HATJoint(rnnt_abstract.AbstractRNNTJoint, Exportable, AdapterModuleMixin):
-    """A Hybrid Autoregressive Transducer Joint Network (HAT Joint Network).
-    A HAT Joint network, comprised of a feedforward model.
-
-    Args:
-        jointnet: A dict-like object which contains the following key-value pairs.
-            encoder_hidden: int specifying the hidden dimension of the encoder net.
-            pred_hidden: int specifying the hidden dimension of the prediction net.
-            joint_hidden: int specifying the hidden dimension of the joint net
-            activation: Activation function used in the joint step. Can be one of
-                ['relu', 'tanh', 'sigmoid'].
-
-            Optionally, it may also contain the following:
-            dropout: float, set to 0.0 by default. Optional dropout applied at the end of the joint net.
-
-        num_classes: int, specifying the vocabulary size that the joint network must predict,
-            excluding the HAT blank token.
-
-        vocabulary: Optional list of strings/tokens that comprise the vocabulary of the joint network.
-            Unused and kept only for easy access for character based encoding HAT models.
-
-        log_softmax: Optional bool, set to None by default. If set as None, will compute the log_softmax()
-            based on the value provided.
-
-        preserve_memory: Optional bool, set to False by default. If the model crashes due to the memory
-            intensive joint step, one might try this flag to empty the tensor cache in pytorch.
-
-            Warning: This will make the forward-backward pass much slower than normal.
-            It also might not fix the OOM if the GPU simply does not have enough memory to compute the joint.
-
-        fuse_loss_wer: Optional bool, set to False by default.
-
-            Fuses the joint forward, loss forward and
-            wer forward steps. In doing so, it trades of speed for memory conservation by creating sub-batches
-            of the provided batch of inputs, and performs Joint forward, loss forward and wer forward (optional),
-            all on sub-batches, then collates results to be exactly equal to results from the entire batch.
-
-            When this flag is set, prior to calling forward, the fields `loss` and `wer` (either one) *must*
-            be set using the `HATJoint.set_loss()` or `HATJoint.set_wer()` methods.
-
-            Further, when this flag is set, the following argument `fused_batch_size` *must* be provided
-            as a non negative integer. This value refers to the size of the sub-batch.
-
-            When the flag is set, the input and output signature of `forward()` of this method changes.
-            Input - in addition to `encoder_outputs` (mandatory argument), the following arguments can be provided.
-                - decoder_outputs (optional). Required if loss computation is required.
-                - encoder_lengths (required)
-                - transcripts (optional). Required for wer calculation.
-                - transcript_lengths (optional). Required for wer calculation.
-                - compute_wer (bool, default false). Whether to compute WER or not for the fused batch.
-
-            Output - instead of the usual `joint` log prob tensor, the following results can be returned.
-                - loss (optional). Returned if decoder_outputs, transcripts and transript_lengths are not None.
-                - wer_numerator + wer_denominator (optional). Returned if transcripts, transcripts_lengths are provided
-                    and compute_wer is set.
-
-        fused_batch_size: Optional int, required if `fuse_loss_wer` flag is set. Determines the size of the
-            sub-batches. Should be any value below the actual batch size per GPU.
-    """
-
-    @property
-    def input_types(self):
-        """Returns definitions of module input ports.
-        """
-        return {
-            "encoder_outputs": NeuralType(('B', 'D', 'T'), AcousticEncodedRepresentation()),
-            "decoder_outputs": NeuralType(('B', 'D', 'T'), EmbeddedTextType()),
-            "encoder_lengths": NeuralType(tuple('B'), LengthsType(), optional=True),
-            "transcripts": NeuralType(('B', 'T'), LabelsType(), optional=True),
-            "transcript_lengths": NeuralType(tuple('B'), LengthsType(), optional=True),
-            "compute_wer": NeuralType(optional=True),
-        }
-
-    @property
-    def output_types(self):
-        """Returns definitions of module output ports.
-        """
-        if not self._fuse_loss_wer:
-            return {
-                "outputs": NeuralType(('B', 'T', 'T', 'D'), LogprobsType()),
-            }
-
-        else:
-            return {
-                "loss": NeuralType(elements_type=LossType(), optional=True),
-                "wer": NeuralType(elements_type=ElementType(), optional=True),
-                "wer_numer": NeuralType(elements_type=ElementType(), optional=True),
-                "wer_denom": NeuralType(elements_type=ElementType(), optional=True),
-            }
-
-    def _prepare_for_export(self, **kwargs):
-        self._fuse_loss_wer = False
-        self.log_softmax = False
-        super()._prepare_for_export(**kwargs)
-
-    def input_example(self, max_batch=1, max_dim=8192):
-        """
-        Generates input examples for tracing etc.
-        Returns:
-            A tuple of input examples.
-        """
-        B, T, U = max_batch, max_dim, max_batch
-        encoder_outputs = torch.randn(B, self.encoder_hidden, T).to(next(self.parameters()).device)
-        decoder_outputs = torch.randn(B, self.pred_hidden, U).to(next(self.parameters()).device)
-        return (encoder_outputs, decoder_outputs)
-
-    @property
-    def disabled_deployment_input_names(self):
-        """Implement this method to return a set of input names disabled for export"""
-        return set(["encoder_lengths", "transcripts", "transcript_lengths", "compute_wer"])
-
-    def __init__(
-        self,
-        jointnet: Dict[str, Any],
-        num_classes: int,
-        num_extra_outputs: int = 0,
-        vocabulary: Optional[List] = None,
-        log_softmax: Optional[bool] = None,
-        preserve_memory: bool = False,
-        fuse_loss_wer: bool = False,
-        fused_batch_size: Optional[int] = None,
-        experimental_fuse_loss_wer: Any = None,
-    ):
-        super().__init__()
-
-        self.vocabulary = vocabulary
-
-        self._vocab_size = num_classes
-        self._num_extra_outputs = num_extra_outputs
-        self._num_classes = num_classes + 1 + num_extra_outputs  # 1 is for blank
-
-        if experimental_fuse_loss_wer is not None:
-            # Override fuse_loss_wer from deprecated argument
-            fuse_loss_wer = experimental_fuse_loss_wer
-
-        self._fuse_loss_wer = fuse_loss_wer
-        self._fused_batch_size = fused_batch_size
-
-        if fuse_loss_wer and (fused_batch_size is None):
-            raise ValueError("If `fuse_loss_wer` is set, then `fused_batch_size` cannot be None!")
-
-        self._loss = None
-        self._wer = None
-
-        # Log softmax should be applied explicitly only for CPU
-        self.log_softmax = log_softmax
-        self.preserve_memory = preserve_memory
-
-        if preserve_memory:
-            logging.warning(
-                "`preserve_memory` was set for the Joint Model. Please be aware this will severely impact "
-                "the forward-backward step time. It also might not solve OOM issues if the GPU simply "
-                "does not have enough memory to compute the joint."
-            )
-
-        # Required arguments
-        self.encoder_hidden = jointnet['encoder_hidden']
-        self.pred_hidden = jointnet['pred_hidden']
-        self.joint_hidden = jointnet['joint_hidden']
-        self.activation = jointnet['activation']
-
-        # Optional arguments
-        dropout = jointnet.get('dropout', 0.0)
-
-        self.pred, self.enc, self.joint_net, self.blank_pred = self._joint_net_modules(
-            num_classes=self._vocab_size,  # non blank symbol
-            pred_n_hidden=self.pred_hidden,
-            enc_n_hidden=self.encoder_hidden,
-            joint_n_hidden=self.joint_hidden,
-            activation=self.activation,
-            dropout=dropout,
-        )
-
-        # Flag needed for RNNT export support
-        self._rnnt_export = False
-
-    @typecheck()
-    def forward(
-        self,
-        encoder_outputs: torch.Tensor,
-        decoder_outputs: Optional[torch.Tensor],
-        encoder_lengths: Optional[torch.Tensor] = None,
-        transcripts: Optional[torch.Tensor] = None,
-        transcript_lengths: Optional[torch.Tensor] = None,
-        compute_wer: bool = False,
-    ) -> Union[torch.Tensor, List[Optional[torch.Tensor]]]:
-        # encoder = (B, D, T)
-        # decoder = (B, D, U) if passed, else None
-        encoder_outputs = encoder_outputs.transpose(1, 2)  # (B, T, D)
-
-        if decoder_outputs is not None:
-            decoder_outputs = decoder_outputs.transpose(1, 2)  # (B, U, D)
-
-        if not self._fuse_loss_wer:
-            if decoder_outputs is None:
-                raise ValueError(
-                    "decoder_outputs passed is None, and `fuse_loss_wer` is not set. "
-                    "decoder_outputs can only be None for fused step!"
-                )
-
-            out, _ = self.joint(encoder_outputs, decoder_outputs)  # [B, T, U, V + 1]
-            return out
-
-        else:
-            # At least the loss module must be supplied during fused joint
-            if self._loss is None or self._wer is None:
-                raise ValueError("`fuse_loss_wer` flag is set, but `loss` and `wer` modules were not provided! ")
-
-            # If fused joint step is required, fused batch size is required as well
-            if self._fused_batch_size is None:
-                raise ValueError("If `fuse_loss_wer` is set, then `fused_batch_size` cannot be None!")
-
-            # When using fused joint step, both encoder and transcript lengths must be provided
-            if (encoder_lengths is None) or (transcript_lengths is None):
-                raise ValueError(
-                    "`fuse_loss_wer` is set, therefore encoder and target lengths " "must be provided as well!"
-                )
-
-            losses = []
-            target_lengths = []
-            batch_size = int(encoder_outputs.size(0))  # actual batch size
-
-            # Iterate over batch using fused_batch_size steps
-            for batch_idx in range(0, batch_size, self._fused_batch_size):
-                begin = batch_idx
-                end = min(begin + self._fused_batch_size, batch_size)
-
-                # Extract the sub batch inputs
-                # sub_enc = encoder_outputs[begin:end, ...]
-                # sub_transcripts = transcripts[begin:end, ...]
-                sub_enc = encoder_outputs.narrow(dim=0, start=begin, length=end - begin)
-                sub_transcripts = transcripts.narrow(dim=0, start=begin, length=end - begin)
-
-                sub_enc_lens = encoder_lengths[begin:end]
-                sub_transcript_lens = transcript_lengths[begin:end]
-
-                # Sub transcripts does not need the full padding of the entire batch
-                # Therefore reduce the decoder time steps to match
-                max_sub_enc_length = sub_enc_lens.max()
-                max_sub_transcript_length = sub_transcript_lens.max()
-
-                if decoder_outputs is not None:
-                    # Reduce encoder length to preserve computation
-                    # Encoder: [sub-batch, T, D] -> [sub-batch, T', D]; T' < T
-                    if sub_enc.shape[1] != max_sub_enc_length:
-                        sub_enc = sub_enc.narrow(dim=1, start=0, length=max_sub_enc_length)
-
-                    # sub_dec = decoder_outputs[begin:end, ...]  # [sub-batch, U, D]
-                    sub_dec = decoder_outputs.narrow(dim=0, start=begin, length=end - begin)  # [sub-batch, U, D]
-
-                    # Reduce decoder length to preserve computation
-                    # Decoder: [sub-batch, U, D] -> [sub-batch, U', D]; U' < U
-                    if sub_dec.shape[1] != max_sub_transcript_length + 1:
-                        sub_dec = sub_dec.narrow(dim=1, start=0, length=max_sub_transcript_length + 1)
-
-                    # Perform joint => [sub-batch, T', U', V + 1]
-                    sub_joint, _ = self.joint(sub_enc, sub_dec)
-
-                    del sub_dec
-
-                    # Reduce transcript length to correct alignment
-                    # Transcript: [sub-batch, L] -> [sub-batch, L']; L' <= L
-                    if sub_transcripts.shape[1] != max_sub_transcript_length:
-                        sub_transcripts = sub_transcripts.narrow(dim=1, start=0, length=max_sub_transcript_length)
-
-                    # Compute sub batch loss
-                    # preserve loss reduction type
-                    loss_reduction = self.loss.reduction
-
-                    # override loss reduction to sum
-                    self.loss.reduction = None
-
-                    # compute and preserve loss
-                    loss_batch = self.loss(
-                        log_probs=sub_joint,
-                        targets=sub_transcripts,
-                        input_lengths=sub_enc_lens,
-                        target_lengths=sub_transcript_lens,
-                    )
-                    losses.append(loss_batch)
-                    target_lengths.append(sub_transcript_lens)
-
-                    # reset loss reduction type
-                    self.loss.reduction = loss_reduction
-
-                else:
-                    losses = None
-
-                # Update WER for sub batch
-                if compute_wer:
-                    sub_enc = sub_enc.transpose(1, 2)  # [B, T, D] -> [B, D, T]
-                    sub_enc = sub_enc.detach()
-                    sub_transcripts = sub_transcripts.detach()
-
-                    # Update WER on each process without syncing
-                    self.wer.update(sub_enc, sub_enc_lens, sub_transcripts, sub_transcript_lens)
-
-                del sub_enc, sub_transcripts, sub_enc_lens, sub_transcript_lens
-
-            # Reduce over sub batches
-            if losses is not None:
-                losses = self.loss.reduce(losses, target_lengths)
-
-            # Collect sub batch wer results
-            if compute_wer:
-                # Sync and all_reduce on all processes, compute global WER
-                wer, wer_num, wer_denom = self.wer.compute()
-                self.wer.reset()
-            else:
-                wer = None
-                wer_num = None
-                wer_denom = None
-
-            return losses, wer, wer_num, wer_denom
-
-    def joint(self, f: torch.Tensor, g: torch.Tensor, return_ilm: bool = False) -> Tuple[torch.Tensor, torch.Tensor]:
-        """
-        Compute the joint step of the network.
-
-        Here,
-        B = Batch size
-        T = Acoustic model timesteps
-        U = Target sequence length
-        H1, H2 = Hidden dimensions of the Encoder / Decoder respectively
-        H = Hidden dimension of the Joint hidden step.
-        V = Vocabulary size of the Decoder (excluding the HAT blank token).
-
-        NOTE:
-            The implementation of this model is slightly modified from the original paper.
-            The original paper proposes the following steps :
-            (enc, dec) -> Expand + Concat + Sum [B, T, U, H1+H2] -> Forward through joint hidden [B, T, U, H] -- *1
-            *1 -> Forward through joint final [B, T, U, V + 1].
-
-            We instead split the joint hidden into joint_hidden_enc and joint_hidden_dec and act as follows:
-            enc -> Forward through joint_hidden_enc -> Expand [B, T, 1, H] -- *1
-            dec -> Forward through joint_hidden_dec -> Expand [B, 1, U, H] -- *2
-            (*1, *2) -> Sum [B, T, U, H] -> Forward through joint final [B, T, U, V + 1].
-
-        Args:
-            f: Output of the Encoder model. A torch.Tensor of shape [B, T, H1]
-            g: Output of the Decoder model. A torch.Tensor of shape [B, U, H2]
-
-        Returns:
-            Logits / log softmaxed tensor of shape (B, T, U, V + 1).
-            Internal LM probabili (B, 1, U, V).
-        """
-        # f = [B, T, H1]
-        f = self.enc(f)
-        f.unsqueeze_(dim=2)  # (B, T, 1, H)
-
-        # g = [B, U, H2]
-        g = self.pred(g)
-        g.unsqueeze_(dim=1)  # (B, 1, U, H)
-
-        inp = f + g  # [B, T, U, H]
-
-        del f
-
-        # Forward adapter modules on joint hidden
-        if self.is_adapter_available():
-            inp = self.forward_enabled_adapters(inp)
-
-        blank_logprob = self.blank_pred(inp)  # [B, T, U, 1]
-        label_logit = self.joint_net(inp)  # [B, T, U, V]
-
-        del inp
-
-        label_logprob = label_logit.log_softmax(dim=-1)
-        scale_prob = torch.clamp(1 - torch.exp(blank_logprob), min=1e-6)
-        label_logprob_scaled = torch.log(scale_prob) + label_logprob  # [B, T, U, V]
-
-        ilm_logprob = None
-        if return_ilm:
-            ilm_logit = self.joint_net(g)  # [B, 1, U, V]
-            ilm_logprob = ilm_logit.log_softmax(dim=-1)
-
-        res = torch.cat((label_logprob_scaled, blank_logprob), dim=-1).contiguous()  # [B, T, U, V+1]
-
-        del g, blank_logprob, label_logprob, label_logit, scale_prob, label_logprob_scaled
-
-        if self.preserve_memory:
-            torch.cuda.empty_cache()
-
-        return res, ilm_logprob
-
-    def _joint_net_modules(self, num_classes, pred_n_hidden, enc_n_hidden, joint_n_hidden, activation, dropout):
-        """
-        Prepare the trainable modules of the Joint Network
-
-        Args:
-            num_classes: Number of output classes (vocab size) excluding the HAT blank token.
-            pred_n_hidden: Hidden size of the prediction network.
-            enc_n_hidden: Hidden size of the encoder network.
-            joint_n_hidden: Hidden size of the joint network.
-            activation: Activation of the joint. Can be one of [relu, tanh, sigmoid]
-            dropout: Dropout value to apply to joint.
-        """
-        pred = torch.nn.Linear(pred_n_hidden, joint_n_hidden)
-        enc = torch.nn.Linear(enc_n_hidden, joint_n_hidden)
-        blank_pred = torch.nn.Sequential(
-            torch.nn.Tanh(), torch.nn.Dropout(p=dropout), torch.nn.Linear(joint_n_hidden, 1), torch.nn.LogSigmoid()
-        )
-
-        if activation not in ['relu', 'sigmoid', 'tanh']:
-            raise ValueError("Unsupported activation for joint step - please pass one of " "[relu, sigmoid, tanh]")
-
-        activation = activation.lower()
-
-        if activation == 'relu':
-            activation = torch.nn.ReLU(inplace=True)
-        elif activation == 'sigmoid':
-            activation = torch.nn.Sigmoid()
-        elif activation == 'tanh':
-            activation = torch.nn.Tanh()
-
-        layers = (
-            [activation]
-            + ([torch.nn.Dropout(p=dropout)] if dropout else [])
-            + [torch.nn.Linear(joint_n_hidden, num_classes)]
-        )
-        return pred, enc, torch.nn.Sequential(*layers), blank_pred
-
-    # Adapter method overrides
-    def add_adapter(self, name: str, cfg: DictConfig):
-        # Update the config with correct input dim
-        cfg = self._update_adapter_cfg_input_dim(cfg)
-        # Add the adapter
-        super().add_adapter(name=name, cfg=cfg)
-
-    def _update_adapter_cfg_input_dim(self, cfg: DictConfig):
-        cfg = adapter_utils.update_adapter_cfg_input_dim(self, cfg, module_dim=self.joint_hidden)
-        return cfg
-
-    @property
-    def num_classes_with_blank(self):
-        return self._num_classes
-
-    @property
-    def num_extra_outputs(self):
-        return self._num_extra_outputs
-
-    @property
-    def loss(self):
-        return self._loss
-
-    def set_loss(self, loss):
-        if not self._fuse_loss_wer:
-            raise ValueError("Attempting to set loss module even though `fuse_loss_wer` is not set!")
-
-        self._loss = loss
-
-    @property
-    def wer(self):
-        return self._wer
-
-    def set_wer(self, wer):
-        if not self._fuse_loss_wer:
-            raise ValueError("Attempting to set WER module even though `fuse_loss_wer` is not set!")
-
-        self._wer = wer
-
-    @property
-    def fuse_loss_wer(self):
-        return self._fuse_loss_wer
-
-    def set_fuse_loss_wer(self, fuse_loss_wer, loss=None, metric=None):
-        self._fuse_loss_wer = fuse_loss_wer
-
-        self._loss = loss
-        self._wer = metric
-
-    @property
-    def fused_batch_size(self):
-        return self._fuse_loss_wer
-
-    def set_fused_batch_size(self, fused_batch_size):
-        self._fused_batch_size = fused_batch_size
